pyproject.toml
[project]
name = "bcpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "paramiko>=3.5.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "scp>=0.15.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
